+++ CHANGELOG.md
@@ -10,6 +10,8 @@ All notable changes to this project will be documented in this file. This projec
 - [#110](https://github.com/laravel-json-api/laravel/issues/110) For requests that modify a relationship, it is now
   possible to get the model or models referenced in the request JSON using the `toOne()` or `toMany()` methods on the
   resource request class.
+- [#113](https://github.com/laravel-json-api/laravel/issues/113) The Eloquent `Number` field can now be configured to
+  accept numeric strings by calling the `acceptStrings()` method on the field.
 
 ## [2.0.0] - 2022-02-09